@@ -9,9 +9,9 @@
     "@testing-library/user-event": "^14.2.0",
     "gh-pages": "^4.0.0",
     "markdown-to-jsx": "^7.1.7",
-    "react": "^18.1.0",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.1.0",
-    "react-dom": "^18.1.0",
+    "react-dom": "^18.2.0",
     "react-router": "^6.4.1",
     "react-router-dom": "^6.4.1",
     "react-scripts": "5.0.1",
