@@ -1,4 +1,4 @@
 export * from './home';
 export * from './blog';
 export * from './music';
-export * from './about';
+export * from './online';
